requirements.txt
aiohttp==3.9.5
google-generativeai==0.8.5
python-dotenv==1.0.1
tenacity==8.5.0
cachetools==5.5.0